Add dependency support
Adding the names of other processes to soft_depends will cause the
current process to wait until after all processes in that list have
started before it itself is started. This is implemented as follows:

During initialization, only processes with no dependencies are started.
When a 'process started' event is received, it triggers a scan over the
list of processes, identifying those that were waiting for this event.
For each waiting process, its full list of dependencies is checked. If
they are all running, the process is started; otherwise it waits for the
next events.

This is not very efficient for large numbers of processes, but that's
likely not a concern here.

While here, make output more readable by adding tabs.